Internet 前后端全部调通

+++ scripts/docker-dev-env.sh
@@ -14,37 +14,31 @@ if [ "net" = $ENV ]; then
   ZMQ_PORT=5557
   WS_PORT=8080
   if [ "bj" = $LOC ]; then
-    KENDS="-kend 10.208.104.9"
-    NSEND="58.213.121.2:10027"
-    ZMQ_OUT="tcp://159.226.41.229:7102"
+    PRO_DIR="/home/zsj/projects/hongbao-ms"
     MAC="02:42:ac:11:00:01"
   else
-    KENDS="-kend 172.16.32.13 -kend 172.16.32.14 -kend 172.16.32.15"
-    NSEND="172.16.32.13:8080"
-    ZMQ_OUT="tcp://58.213.121.2:10025"
+    PRO_DIR="/root/projects/hongbao-ms"
     MAC="02:42:ac:11:00:02"
   fi
 else
   ZMQ_PORT=5558
   WS_PORT=8080
-  KENDS=""
+  PRO_DIR="/root/projects/hongbao-ms"
 fi
 
 docker pull registry.cn-beijing.aliyuncs.com/zhengsj/hongbao:msd
-docker rmi $(docker image ls -f dangling=true -q)
-docker stop msd
-docker rm msd
+if [ "" != "$(docker image ls -f dangling=true -q)" ]; then
+  docker rmi $(docker image ls -f dangling=true -q)
+fi
+docker stop msd-dev
+docker rm msd-dev
+
 docker run -it \
-  --name msd \
+  --name msd-dev \
   -p $WS_PORT:5554 \
   -p $ZMQ_PORT:5553 \
   -p 5552:5552 \
+  -v $PRO_DIR:/hongbao-ms \
   --mac-address $MAC \
   registry.cn-beijing.aliyuncs.com/zhengsj/hongbao:msd \
-  msd \
-    -ws   0.0.0.0:5554 \
-    -zmq  tcp://0.0.0.0:5553 \
-    -log  0.0.0.0:5552 \
-    -nsend $NSEND \
-    -zmq-out $ZMQ_OUT \
-    -$ENV $KENDS
+  bash
